Разбей скрипт на модули и подключи их
